requirements.txt
pillow
numpy
tqdm
SimpleITK
h5py
tensorboard
opencv-python
torchsummary